nostr-dav: fix aarch64 -Werror=format in load_pending_tombstones
sqlite3_column_int64() returns sqlite3_int64 (== long long int always),
but G_GINT64_FORMAT expands to "li" on 64-bit Linux (where gint64 is
long int). On x86_64 the sizes match so gcc doesn't warn; on aarch64
gcc's -Werror=format catches the typedef mismatch and the build
aborts at [596/609].

Fix: cast sqlite3_int64 -> gint64 explicitly. Both are guaranteed
64-bit signed; the cast just aligns to G_GINT64_FORMAT's expected type.

Discovered by nostrc-qngf (Debian installed-unit acceptance on
aarch64 lab). Introduced by e8eafde7 (Wave 4 tombstone landing).

+++ gnome/nostr-dav/src/nd-publisher.c
@@ -1027,9 +1027,13 @@ load_pending_tombstones(NdPublisher *self, gint64 now_ts)
     NdOutboxRow *row = g_new0(NdOutboxRow, 1);
     row->outbox = ND_OUTBOX_TOMBSTONE;
     row->col    = ND_STORE_COLLECTION_EVENTS; /* unused for tombstones */
-    /* Stringify the AUTOINCREMENT id so it fits the row_id contract. */
+    /* Stringify the AUTOINCREMENT id so it fits the row_id contract.
+     * Cast sqlite3_int64 -> gint64 because on some 64-bit platforms
+     * (e.g. aarch64 Linux) sqlite3_int64 is `long long int` while gint64
+     * is `long int`, and G_GINT64_FORMAT ("li") triggers -Werror=format
+     * without the cast. Both are guaranteed 64-bit signed. */
     row->row_id = g_strdup_printf("%" G_GINT64_FORMAT,
-                                  sqlite3_column_int64(stmt, 0));
+                                  (gint64)sqlite3_column_int64(stmt, 0));
     const unsigned char *sj = sqlite3_column_text(stmt, 1);
     row->signed_json = sj ? g_strdup((const gchar *)sj) : NULL;
     const unsigned char *tj = sqlite3_column_text(stmt, 2);
